IOT BASED SMART SOCIAL DISTANCING REMAINDER
During the time of required social distancing, this project measures how close you are to others for safety purpose. It detects how you are closer and the distance between one people to another and it also measures the temperature of the people and if you are getting close and the temperature is high,  it gives a alarm message and the LCD flashes.so that you can identify and detect that you are close it is not safe. It could identify easily and protect us from virus transmission. It is highly beneficial for us during lockdown period .All high-risk patients are tracked easily using the internet-based network. This technology is used for biometric measurements like blood pressure, heartbeat and glucose level too not only for temperature.